moved instanceTile from base to molecules
