+++ src/components/molecules/InstanceTile.unit.js
@@ -1,5 +1,5 @@
 import InstanceTile from "./InstanceTile";
 
-describe("@components/base/InstanceTile", () => {
+describe("@components/molecules/InstanceTile", () => {
 	it(...isValidComponent(InstanceTile));
 });
